refactor to generalize (messy & untested)
