+++ src/box_training_methods/enums.py
